+++ config.js
@@ -1,0 +1,24 @@
+module.exports = {
+  config: {
+    'blogName': '지나가던 개발자',
+    'name': 'Kim Songhyun',
+    'description': '카카오 프론트엔드 개발자 since 2018. 01.',
+    'host': 'https://znagadeon.dev',
+    'googleSearch': '981b407e9acf0f1f0',
+    'googleAnalytics': 'G-TP8Y4MDDCD',
+    'links': {
+      'rss': '/rss.xml',
+      'github': 'znagadeon',
+      'linkedin': 'kim-songhyun',
+      'twitter': 'znagadeon',
+      'gravatar': 'c1db11b57685a2fc71655a869cdac430',
+      'donation': 'https://toss.me/znagadeon',
+    },
+    'comment': {
+      'repository': 'znagadeon/homepage',
+      'repoId': 'MDEwOlJlcG9zaXRvcnkyMzc0ODA5OTk=',
+      'category': 'General',
+      'categoryId': 'MDE4OkRpc2N1c3Npb25DYXRlZ29yeTMyMzIwODgy',
+    },
+  },
+};
